:zap: Refactor: combine apis to postAPI
Refactor: combine apis to postAPI
